Add unit tests for parseValuesToCreatePool function
