+++ src/hooks/aelin/useAelinCreateUpFrontDeal.tsx
@@ -5,7 +5,6 @@ import styled from 'styled-components'
 import { isAddress } from '@ethersproject/address'
 import { BigNumber } from '@ethersproject/bignumber'
 import { BigNumberish } from '@ethersproject/bignumber'
-import { HashZero, MaxUint256 } from '@ethersproject/constants'
 import { parseUnits } from '@ethersproject/units'
 import { wei } from '@synthetixio/wei'
 
@@ -16,7 +15,7 @@ import { AddressWhitelistProps } from '@/src/components/pools/whitelist/addresse
 import { NftType } from '@/src/components/pools/whitelist/nft/nftWhiteListReducer'
 import { ChainsValues, getKeyChainByValue } from '@/src/constants/chains'
 import { contracts } from '@/src/constants/contracts'
-import { BASE_DECIMALS, EXCHANGE_DECIMALS, ZERO_ADDRESS, ZERO_BN } from '@/src/constants/misc'
+import { BASE_DECIMALS, EXCHANGE_DECIMALS, ZERO_ADDRESS } from '@/src/constants/misc'
 import { Privacy } from '@/src/constants/pool'
 import { Token } from '@/src/constants/token'
 import {
@@ -25,6 +24,7 @@ import {
 } from '@/src/hooks/contracts/useAelinDirectDealCreateTransaction'
 import { GasOptions, useTransactionModal } from '@/src/providers/transactionModalProvider'
 import { useWeb3Connection } from '@/src/providers/web3ConnectionProvider'
+import { parseValuesToCreateUpFrontDeal } from '@/src/utils/aelinPoolUtils'
 import { getDuration, isEmptyDuration, secondsToDhm } from '@/src/utils/date'
 import { formatNumber } from '@/src/utils/formatNumber'
 import { mergeArrayKeyValuePairs } from '@/src/utils/mergeArrayKeyValuePairs'
@@ -401,137 +401,6 @@ export const getCreateDealSummaryData = (
     value: step.getSummaryValue(createDealState),
   }))
 
-const parseValuesToCreateUpFrontDeal = (
-  createDealState: CreateUpFrontDealStateComplete,
-  sponsor: string,
-  merkleRoot?: string,
-  ipfsHash?: string,
-): CreateUpFrontDealValues => {
-  const {
-    dealAttributes,
-    dealPrivacy,
-    dealToken,
-    exchangeRates,
-    holderAddress,
-    investmentToken,
-    redemptionDeadline,
-    sponsorFee,
-    vestingSchedule,
-    whitelist,
-    withMerkleTree,
-  } = createDealState
-  const now = new Date()
-
-  const underlyingDealTokenTotal = wei(
-    exchangeRates?.investmentTokenToRaise,
-    dealToken.decimals,
-  ).mul(wei(exchangeRates.exchangeRates, dealToken.decimals))
-
-  const exchangeRatesInWei = wei(exchangeRates.exchangeRates, investmentToken.decimals)
-
-  const dealTokenTotalInWei = wei(
-    exchangeRates.investmentTokenToRaise,
-    investmentToken.decimals,
-  ).mul(exchangeRatesInWei)
-
-  const investmentPerDeal = wei(exchangeRates.investmentTokenToRaise, investmentToken.decimals).div(
-    dealTokenTotalInWei,
-  )
-
-  const purchaseRaiseMinimum = exchangeRates?.minimumAmount
-    ? parseUnits(exchangeRates.minimumAmount.toString(), investmentToken.decimals)
-    : ZERO_BN
-
-  const redemptionDeadlineDuration = getDuration(
-    now,
-    redemptionDeadline.days,
-    redemptionDeadline.hours,
-    redemptionDeadline.minutes,
-  )
-
-  const vestingCliffDuration = getDuration(
-    now,
-    vestingSchedule.vestingCliff?.days as number,
-    vestingSchedule.vestingCliff?.hours as number,
-    vestingSchedule.vestingCliff?.minutes as number,
-  )
-
-  const vestingPeriodDuration = getDuration(
-    now,
-    vestingSchedule.vestingPeriod?.days as number,
-    vestingSchedule.vestingPeriod?.hours as number,
-    vestingSchedule.vestingPeriod?.minutes as number,
-  )
-
-  let dealAddresses: string[] = []
-  let dealAddressesAmounts: BigNumberish[] = []
-  let nftCollectionRules: NftCollectionRulesProps[] = []
-
-  if (
-    dealPrivacy === Privacy.PRIVATE &&
-    [NftType.erc1155, NftType.erc721].some((type) => !Object.hasOwn(createDealState, type))
-  ) {
-    const formattedWhiteList = whitelist.reduce((accum, curr) => {
-      const { address, amount } = curr
-
-      if (!isAddress(address)) return accum
-
-      if (withMerkleTree) {
-        accum.push({
-          address,
-          amount: amount ? amount : MaxUint256.toNumber(),
-        })
-      } else {
-        accum.push({
-          address,
-          amount: amount ? amount : MaxUint256.toNumber(),
-        })
-      }
-
-      return accum
-    }, [] as { address: string; amount: BigNumberish }[])
-
-    dealAddresses = formattedWhiteList.map(({ address }) => address)
-    dealAddressesAmounts = formattedWhiteList.map(({ amount }) => amount)
-  }
-
-  if (dealPrivacy === Privacy.NFT && Object.hasOwn(createDealState, NftType.erc721)) {
-    nftCollectionRules = [...createDealState[NftType.erc721]]
-  }
-
-  if (dealPrivacy === Privacy.NFT && Object.hasOwn(createDealState, NftType.erc1155)) {
-    nftCollectionRules = [...createDealState[NftType.erc1155]]
-  }
-
-  return [
-    {
-      name: dealAttributes.name,
-      symbol: dealAttributes.symbol,
-      purchaseToken: investmentToken.address,
-      underlyingDealToken: dealToken.address,
-      holder: holderAddress,
-      sponsor,
-      sponsorFee: sponsorFee ? parseUnits(sponsorFee?.toString(), BASE_DECIMALS) : ZERO_BN,
-      merkleRoot: merkleRoot ?? HashZero,
-      ipfsHash: ipfsHash ?? HashZero,
-    },
-    {
-      underlyingDealTokenTotal: underlyingDealTokenTotal.toBN(),
-      purchaseTokenPerDealToken: investmentPerDeal.toBN(),
-      purchaseRaiseMinimum,
-      purchaseDuration: redemptionDeadlineDuration,
-      vestingPeriod: vestingPeriodDuration,
-      vestingCliffPeriod: vestingCliffDuration,
-      allowDeallocation: !exchangeRates.isCapped,
-    },
-    nftCollectionRules,
-    {
-      allowListAddresses: dealAddresses,
-      allowListAmounts: dealAddressesAmounts,
-    },
-  ]
-}
-
 const initialState: CreateUpFrontDealState = {
   [CreateUpFrontDealSteps.dealAttributes]: {
     name: '',
@@ -675,6 +544,15 @@ export default function useAelinCreateDeal(chainId: ChainsValues) {
         createDealState as CreateUpFrontDealStateComplete,
         address ?? ZERO_ADDRESS,
       )
+
+    console.log('createDealState: ', createDealState)
+
+    const x = await parseValuesToCreateUpFrontDeal(
+      createDealState as CreateUpFrontDealStateComplete,
+      address ?? ZERO_ADDRESS,
+    )
+
+    console.log('x: ', x)
 
     const formattedNftCollectionRules = nftCollectionRules.map((collection) => ({
       ...collection,
